Refactoring for terminal API utility methods
