+++ adyenv6core/src/com/adyen/v6/facades/DefaultAdyenCheckoutFacade.java
@@ -20,6 +20,28 @@
  */
 package com.adyen.v6.facades;
 
+import java.io.IOException;
+import java.security.SignatureException;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import java.util.SortedMap;
+import java.util.TreeMap;
+import java.util.UUID;
+import java.util.concurrent.TimeUnit;
+import java.util.stream.Collectors;
+import javax.annotation.Resource;
+import javax.servlet.http.HttpServletRequest;
+import org.apache.commons.lang3.StringUtils;
+import org.apache.commons.lang3.exception.ExceptionUtils;
+import org.apache.log4j.Logger;
+import org.springframework.ui.Model;
+import org.springframework.util.Assert;
+import org.springframework.validation.BindingResult;
 import com.adyen.Util.HMACValidator;
 import com.adyen.Util.Util;
 import com.adyen.constants.HPPConstants;
@@ -28,10 +50,7 @@ import com.adyen.model.Card;
 import com.adyen.model.PaymentResult;
 import com.adyen.model.checkout.PaymentMethod;
 import com.adyen.model.checkout.PaymentsResponse;
-import com.adyen.model.nexo.DocumentQualifierType;
 import com.adyen.model.nexo.ErrorConditionType;
-import com.adyen.model.nexo.OutputText;
-import com.adyen.model.nexo.PaymentReceipt;
 import com.adyen.model.nexo.ResultType;
 import com.adyen.model.recurring.Recurring;
 import com.adyen.model.recurring.RecurringDetail;
@@ -49,7 +68,7 @@ import com.adyen.v6.repository.OrderRepository;
 import com.adyen.v6.service.AdyenOrderService;
 import com.adyen.v6.service.AdyenPaymentService;
 import com.adyen.v6.service.AdyenTransactionService;
-import com.google.common.base.Splitter;
+import com.adyen.v6.util.TerminalAPIUtil;
 import com.google.gson.Gson;
 import de.hybris.platform.commercefacades.i18n.I18NFacade;
 import de.hybris.platform.commercefacades.order.CheckoutFacade;
@@ -78,31 +97,6 @@ import de.hybris.platform.servicelayer.search.FlexibleSearchService;
 import de.hybris.platform.servicelayer.session.SessionService;
 import de.hybris.platform.store.BaseStoreModel;
 import de.hybris.platform.store.services.BaseStoreService;
-import org.apache.commons.lang3.StringUtils;
-import org.apache.commons.lang3.exception.ExceptionUtils;
-import org.apache.log4j.Logger;
-import org.springframework.ui.Model;
-import org.springframework.util.Assert;
-import org.springframework.validation.BindingResult;
-
-import javax.annotation.Resource;
-import javax.servlet.http.HttpServletRequest;
-import java.io.IOException;
-import java.io.UnsupportedEncodingException;
-import java.security.SignatureException;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.SortedMap;
-import java.util.TreeMap;
-import java.util.UUID;
-import java.util.concurrent.TimeUnit;
-import java.util.stream.Collectors;
-
 import static com.adyen.constants.ApiConstants.Redirect.Data.MD;
 import static com.adyen.constants.ApiConstants.ThreeDS2Property.CHALLENGE_RESULT;
 import static com.adyen.constants.ApiConstants.ThreeDS2Property.FINGERPRINT_RESULT;
@@ -697,7 +691,7 @@ public class DefaultAdyenCheckoutFacade implements AdyenCheckoutFacade {
         orderData.setAdyenMultibancoReference(paymentsResponse.getMultibancoReference());
 
         orderData.setAdyenPosReceipt(paymentsResponse.getAdditionalData().get("pos.receipt"));
-        
+
         return orderData;
     }
 
@@ -836,6 +830,7 @@ public class DefaultAdyenCheckoutFacade implements AdyenCheckoutFacade {
         if (baseStore.getAdyenBoleto() == null || ! baseStore.getAdyenBoleto()) {
             return false;
         }
+
         CartData cartData = getCheckoutFacade().getCheckoutCart();
         String currency = cartData.getTotalPrice().getCurrencyIso();
         String country = cartData.getDeliveryAddress().getCountry().getIsocode();
@@ -857,6 +852,7 @@ public class DefaultAdyenCheckoutFacade implements AdyenCheckoutFacade {
     @Override
     public boolean showRememberDetails() {
         BaseStoreModel baseStore = baseStoreService.getCurrentBaseStore();
+
         /*
          * The show remember me checkout should only be shown as the
          * user is logged in and the recurirng mode is set to ONECLICK or ONECLICK,RECURRING
@@ -867,6 +863,7 @@ public class DefaultAdyenCheckoutFacade implements AdyenCheckoutFacade {
                 return true;
             }
         }
+
         return false;
     }
 
@@ -1085,7 +1082,7 @@ public class DefaultAdyenCheckoutFacade implements AdyenCheckoutFacade {
 
         String serviceId = request.getAttribute("originalServiceId").toString();
         TerminalAPIResponse terminalApiResponse = getAdyenPaymentService().sendSyncPosPaymentRequest(cartData, customer, serviceId);
-        ResultType resultType = getPaymentResult(terminalApiResponse);
+        ResultType resultType = TerminalAPIUtil.getPaymentResult(terminalApiResponse);
 
         if (ResultType.SUCCESS == resultType) {
             PaymentsResponse paymentsResponse = getPosPaymentResponseConverter().convert(terminalApiResponse.getSaleToPOIResponse());
@@ -1102,30 +1099,23 @@ public class DefaultAdyenCheckoutFacade implements AdyenCheckoutFacade {
 
         String originalServiceId = request.getAttribute("originalServiceId").toString();
         TerminalAPIResponse terminalApiResponse = getAdyenPaymentService().sendSyncPosStatusRequest(cartData, originalServiceId);
-        ResultType statusResult = getStatusResult(terminalApiResponse);
+        ResultType statusResult = TerminalAPIUtil.getStatusResult(terminalApiResponse);
 
         if (statusResult != null) {
             if (statusResult == ResultType.SUCCESS) {
                 //this will be success even if payment is failed. because this belongs to status call not payment call
-                ResultType paymentResult = getPaymentResult(terminalApiResponse);
+                ResultType paymentResult = TerminalAPIUtil.getPaymentResult(terminalApiResponse);
                 if (paymentResult == ResultType.SUCCESS) {
                     PaymentsResponse paymentsResponse = getPosPaymentResponseConverter().convert(terminalApiResponse.getSaleToPOIResponse());
                     if (terminalApiResponse.getSaleToPOIResponse().getTransactionStatusResponse().getRepeatedMessageResponse().getRepeatedResponseMessageBody().getPaymentResponse().getPaymentReceipt()
                             != null) {
-                        LOGGER.debug("receipt is present ==================" + terminalApiResponse.getSaleToPOIResponse()
-                                                                                                  .getTransactionStatusResponse()
-                                                                                                  .getRepeatedMessageResponse()
-                                                                                                  .getRepeatedResponseMessageBody()
-                                                                                                  .getPaymentResponse()
-                                                                                                  .getPaymentReceipt());
-                        String posReceipt = formatTerminalAPIReceipt(terminalApiResponse.getSaleToPOIResponse()
-                                                                                        .getTransactionStatusResponse()
-                                                                                        .getRepeatedMessageResponse()
-                                                                                        .getRepeatedResponseMessageBody()
-                                                                                        .getPaymentResponse()
-                                                                                        .getPaymentReceipt());
+                        String posReceipt = TerminalAPIUtil.formatTerminalAPIReceipt(terminalApiResponse.getSaleToPOIResponse()
+                                                                                                        .getTransactionStatusResponse()
+                                                                                                        .getRepeatedMessageResponse()
+                                                                                                        .getRepeatedResponseMessageBody()
+                                                                                                        .getPaymentResponse()
+                                                                                                        .getPaymentReceipt());
                         if (StringUtils.isNotEmpty(posReceipt)) {
-
                             paymentsResponse.putAdditionalDataItem("pos.receipt", posReceipt);
                         }
 
@@ -1134,9 +1124,8 @@ public class DefaultAdyenCheckoutFacade implements AdyenCheckoutFacade {
                 } else {
                     throw new AdyenNonAuthorizedPaymentException(terminalApiResponse);
                 }
-
             } else {
-                ErrorConditionType errorCondition = getErrorConditionForStatus(terminalApiResponse);
+                ErrorConditionType errorCondition = TerminalAPIUtil.getErrorConditionForStatus(terminalApiResponse);
                 //If transaction is still in progress, keep retrying in 5 seconds.
                 if (errorCondition == ErrorConditionType.IN_PROGRESS) {
                     TimeUnit.SECONDS.sleep(5);
@@ -1156,83 +1145,6 @@ public class DefaultAdyenCheckoutFacade implements AdyenCheckoutFacade {
             }
         }
         return null;
-    }
-
-    public String formatTerminalAPIReceipt(List<PaymentReceipt> paymentReceipts) {
-        String formattedHtml = "<table class='terminal-api-receipt'>";
-        for (PaymentReceipt paymentReceipt : paymentReceipts) {
-            if (paymentReceipt.getDocumentQualifier() == DocumentQualifierType.CUSTOMER_RECEIPT) {
-                for (OutputText outputText : paymentReceipt.getOutputContent().getOutputText()) {
-                    String test = outputText.getText();
-                    Map<String, String> map = Splitter.on("&").withKeyValueSeparator('=').split(test);
-                    formattedHtml += "<tr class='terminal-api-receipt'>";
-                    if ((map.get("name") != null)) {
-                        formattedHtml += "<td class='terminal-api-receipt-name'>" + map.get("name") + "</td>";
-                    } else {
-                        formattedHtml += "<td class='terminal-api-receipt-name'>&nbsp;</td>";
-                    }
-                    if (map.get("value") != null) {
-                        formattedHtml += "<td class='terminal-api-receipt-value' align='right'>" + map.get("value") + "</td>";
-                    } else {
-                        formattedHtml += "<td class='terminal-api-receipt-value' align='right'>&nbsp;</td>";
-                    }
-                    formattedHtml += "</tr>";
-                }
-            }
-        }
-        formattedHtml += "</table>";
-        try {
-            formattedHtml = java.net.URLDecoder.decode(formattedHtml, "UTF-8");
-        } catch (UnsupportedEncodingException e) {
-            e.printStackTrace();
-        }
-        return formattedHtml;
-    }
-
-
-    /**
-     * @param terminalApiResponse
-     * @return Result from payment response present in paymentResponse in terminalApiResponse for POS payment or POS status call, if present.
-     * Otherwise returns Failure.
-     */
-    public static ResultType getPaymentResult(TerminalAPIResponse terminalApiResponse) {
-
-        if (terminalApiResponse != null && terminalApiResponse.getSaleToPOIResponse() != null) {
-            if (terminalApiResponse.getSaleToPOIResponse().getPaymentResponse() != null) {
-                return terminalApiResponse.getSaleToPOIResponse().getPaymentResponse().getResponse().getResult();
-            } else if (terminalApiResponse.getSaleToPOIResponse().getTransactionStatusResponse() != null) {
-                return terminalApiResponse.getSaleToPOIResponse()
-                                          .getTransactionStatusResponse()
-                                          .getRepeatedMessageResponse()
-                                          .getRepeatedResponseMessageBody()
-                                          .getPaymentResponse()
-                                          .getResponse()
-                                          .getResult();
-            }
-        }
-        return ResultType.FAILURE;
-    }
-
-    public static ResultType getStatusResult(TerminalAPIResponse terminalApiResponse) {
-        if (terminalApiResponse.getSaleToPOIResponse() != null && terminalApiResponse.getSaleToPOIResponse().getTransactionStatusResponse() != null) {
-            return terminalApiResponse.getSaleToPOIResponse().getTransactionStatusResponse().getResponse().getResult();
-        }
-        return null;
-
-    }
-
-    private ErrorConditionType getErrorConditionForPayment(TerminalAPIResponse terminalApiResponse) {
-        return terminalApiResponse.getSaleToPOIResponse()
-                                  .getTransactionStatusResponse()
-                                  .getRepeatedMessageResponse()
-                                  .getRepeatedResponseMessageBody()
-                                  .getPaymentResponse()
-                                  .getResponse()
-                                  .getErrorCondition();
-    }
-
-    private ErrorConditionType getErrorConditionForStatus(TerminalAPIResponse terminalApiResponse) {
-        return terminalApiResponse.getSaleToPOIResponse().getTransactionStatusResponse().getResponse().getErrorCondition();
     }
